+++ src/pages/personal/contact-details/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import FormFooter from "../../components/students/FormFooter";
+import FormFooter from "../../../components/students/FormFooter";
 
 const index = () => {
   return (
@@ -148,7 +148,10 @@ const index = () => {
           </form>
         </div>
       </div>
-      <FormFooter />
+      <FormFooter
+        forwardLink="/personal/passport-details"
+        backLink="/personal/about-student"
+      />
     </div>
   );
 };
